+++ services/rollout-service/pkg/argo/argo_test.go
@@ -19,6 +19,7 @@ package argo
 import (
 	"context"
 	"fmt"
+	"github.com/google/go-cmp/cmp"
 	"io"
 	"k8s.io/apimachinery/pkg/watch"
 	"testing"
@@ -97,21 +98,29 @@ func (a *mockArgoProcessor) checkEvent(ev *v1alpha1.ApplicationWatchEvent) bool 
 	return false
 }
 
-func (a *mockArgoProcessor) Consume(t *testing.T, ctx context.Context, expectedTypes []string) error {
+func (a *mockArgoProcessor) Consume(t *testing.T, ctx context.Context, expectedTypes []string, existingArgoApps bool) error {
 	appsKnownToArgo := map[string]map[string]*v1alpha1.Application{}
+	envAppsKnownToArgo := make(map[string]*v1alpha1.Application)
+
 	for {
 		select {
 		case overview := <-a.trigger:
 			for _, envGroup := range overview.EnvironmentGroups {
 				for _, env := range envGroup.Environments {
-					envAppsKnownToArgo := appsKnownToArgo[env.Name]
-					err := a.DeleteArgoApps(ctx, envAppsKnownToArgo, env.Applications)
+					if ok := appsKnownToArgo[env.Name]; ok != nil {
+						envAppsKnownToArgo = appsKnownToArgo[env.Name]
+						err := a.DeleteArgoApps(ctx, envAppsKnownToArgo, env.Applications)
 
-					if err != nil {
-						continue
+						if err != nil {
+							continue
+						}
 					}
 
 					for _, app := range env.Applications {
+						if existingArgoApps {
+							envAppsKnownToArgo[app.Name] = CreateArgoApplication(overview, app, env)
+							appsKnownToArgo[env.Name] = envAppsKnownToArgo
+						}
 						a.CreateOrUpdateApp(ctx, overview, app, env, envAppsKnownToArgo)
 					}
 				}
@@ -184,9 +193,9 @@ func (m *mockApplicationServiceClient) Delete(ctx context.Context, req *applicat
 	}
 }
 
-func (m *mockApplicationServiceClient) UpdateSpec(ctx context.Context, req *application.ApplicationUpdateSpecRequest) {
+func (m *mockApplicationServiceClient) Update(ctx context.Context, req *application.ApplicationUpdateRequest) {
 	for _, a := range m.Apps {
-		if a.App.Name == *req.Name {
+		if a.App.Name == req.Application.Name {
 			updateApp := &ArgoApp{App: a.App, LastEvent: "MODIFIED"}
 			m.Apps = append(m.Apps, updateApp)
 			break
@@ -231,9 +240,9 @@ func TestArgoConsume(t *testing.T) {
 		Steps                 []step
 		Overview              *api.GetOverviewResponse
 		ExpectedError         string
-		ExpectedReady         bool
 		ExpectedConsumed      int
 		ExpectedConsumedTypes []string
+		ExistingArgoApps      bool
 	}{
 		{
 			Name: "when ctx in cancelled no app is processed",
@@ -286,7 +295,6 @@ func TestArgoConsume(t *testing.T) {
 				},
 				GitRevision: "1234",
 			},
-			ExpectedReady: false,
 		},
 		{
 			Name: "an error is detected",
@@ -342,12 +350,11 @@ func TestArgoConsume(t *testing.T) {
 				},
 				GitRevision: "1234",
 			},
-			ExpectedReady:    false,
 			ExpectedError:    "watching applications: no",
 			ExpectedConsumed: 1,
 		},
 		{
-			Name: "create an app and update it",
+			Name: "create an app",
 			Steps: []step{
 				{
 					Event: &v1alpha1.ApplicationWatchEvent{
@@ -367,6 +374,60 @@ func TestArgoConsume(t *testing.T) {
 						},
 					},
 				},
+				{
+					RecvErr:       status.Error(codes.Canceled, "context cancelled"),
+					CancelContext: true,
+				},
+			},
+			Overview: &api.GetOverviewResponse{
+				Applications: map[string]*api.Application{
+					"foo": {
+						Releases: []*api.Release{
+							{
+								Version:        1,
+								SourceCommitId: "00001",
+							},
+						},
+						Team: "footeam",
+					},
+				},
+				EnvironmentGroups: []*api.EnvironmentGroup{
+					{
+
+						EnvironmentGroupName: "staging-group",
+						Environments: []*api.Environment{
+							{
+								Name: "staging",
+								Applications: map[string]*api.Environment_Application{
+									"foo": {
+										Name:    "foo",
+										Version: 1,
+										DeploymentMetaData: &api.Environment_Application_DeploymentMetaData{
+											DeployTime: "123456789",
+										},
+									},
+								},
+								Priority: api.Priority_UPSTREAM,
+								Config: &api.EnvironmentConfig{
+									Argocd: &api.EnvironmentConfig_ArgoCD{
+										Destination: &api.EnvironmentConfig_ArgoCD_Destination{
+											Name:   "staging",
+											Server: "test-server",
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+				GitRevision: "1234",
+			},
+			ExpectedConsumedTypes: []string{"ADDED"},
+			ExpectedConsumed:      2,
+		},
+		{
+			Name: "updates an already existing app",
+			Steps: []step{
 				{
 					Event: &v1alpha1.ApplicationWatchEvent{
 						Type: "MODIFIED",
@@ -433,34 +494,16 @@ func TestArgoConsume(t *testing.T) {
 				},
 				GitRevision: "1234",
 			},
-			ExpectedReady:         true,
-			ExpectedConsumedTypes: []string{"ADDED", "MODIFIED"},
-			ExpectedConsumed:      2,
+			ExpectedConsumed:      1,
+			ExpectedConsumedTypes: []string{"MODIFIED"},
+			ExistingArgoApps:      true,
 		},
 		{
-			Name: "create an app and try to create it again",
+			Name: "two applications in the overview but only one is updated",
 			Steps: []step{
 				{
 					Event: &v1alpha1.ApplicationWatchEvent{
-						Type: "ADDED",
-						Application: v1alpha1.Application{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:        "foo",
-								Annotations: map[string]string{},
-							},
-							Spec: v1alpha1.ApplicationSpec{
-								Project: "",
-							},
-							Status: v1alpha1.ApplicationStatus{
-								Sync:   v1alpha1.SyncStatus{Revision: "1234"},
-								Health: v1alpha1.HealthStatus{},
-							},
-						},
-					},
-				},
-				{
-					Event: &v1alpha1.ApplicationWatchEvent{
-						Type: "ADDED",
+						Type: "MODIFIED",
 						Application: v1alpha1.Application{
 							ObjectMeta: metav1.ObjectMeta{
 								Name:        "foo",
@@ -492,93 +535,11 @@ func TestArgoConsume(t *testing.T) {
 						},
 						Team: "footeam",
 					},
-				},
-				EnvironmentGroups: []*api.EnvironmentGroup{
-					{
-
-						EnvironmentGroupName: "staging-group",
-						Environments: []*api.Environment{
-							{
-								Name: "staging",
-								Applications: map[string]*api.Environment_Application{
-									"foo": {
-										Name:    "foo",
-										Version: 1,
-										DeploymentMetaData: &api.Environment_Application_DeploymentMetaData{
-											DeployTime: "123456789",
-										},
-									},
-								},
-								Priority: api.Priority_UPSTREAM,
-								Config: &api.EnvironmentConfig{
-									Argocd: &api.EnvironmentConfig_ArgoCD{
-										Destination: &api.EnvironmentConfig_ArgoCD_Destination{
-											Name:   "staging",
-											Server: "test-server",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-				GitRevision: "1234",
-			},
-			ExpectedReady:         true,
-			ExpectedConsumed:      1,
-			ExpectedConsumedTypes: []string{"ADDED"},
-		},
-		{
-			Name: "update a non-existing app and try to create it",
-			Steps: []step{
-				{
-					Event: &v1alpha1.ApplicationWatchEvent{
-						Type: "UPDATED",
-						Application: v1alpha1.Application{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:        "foo",
-								Annotations: map[string]string{},
-							},
-							Spec: v1alpha1.ApplicationSpec{
-								Project: "",
-							},
-							Status: v1alpha1.ApplicationStatus{
-								Sync:   v1alpha1.SyncStatus{Revision: "1234"},
-								Health: v1alpha1.HealthStatus{},
-							},
-						},
-					},
-				},
-				{
-					Event: &v1alpha1.ApplicationWatchEvent{
-						Type: "ADDED",
-						Application: v1alpha1.Application{
-							ObjectMeta: metav1.ObjectMeta{
-								Name:        "foo",
-								Annotations: map[string]string{},
-							},
-							Spec: v1alpha1.ApplicationSpec{
-								Project: "",
-							},
-							Status: v1alpha1.ApplicationStatus{
-								Sync:   v1alpha1.SyncStatus{Revision: "1234"},
-								Health: v1alpha1.HealthStatus{},
-							},
-						},
-					},
-				},
-				{
-					RecvErr:       status.Error(codes.Canceled, "context cancelled"),
-					CancelContext: true,
-				},
-			},
-			Overview: &api.GetOverviewResponse{
-				Applications: map[string]*api.Application{
-					"foo": {
+					"foo2": {
 						Releases: []*api.Release{
 							{
 								Version:        1,
-								SourceCommitId: "00001",
+								SourceCommitId: "00012",
 							},
 						},
 						Team: "footeam",
@@ -599,6 +560,13 @@ func TestArgoConsume(t *testing.T) {
 											DeployTime: "123456789",
 										},
 									},
+									"foo2": {
+										Name:    "foo2",
+										Version: 1,
+										DeploymentMetaData: &api.Environment_Application_DeploymentMetaData{
+											DeployTime: "1234567892",
+										},
+									},
 								},
 								Priority: api.Priority_UPSTREAM,
 								Config: &api.EnvironmentConfig{
@@ -615,9 +583,80 @@ func TestArgoConsume(t *testing.T) {
 				},
 				GitRevision: "1234",
 			},
-			ExpectedReady:         true,
 			ExpectedConsumed:      1,
-			ExpectedConsumedTypes: []string{"ADDED"},
+			ExpectedConsumedTypes: []string{"MODIFIED"},
+			ExistingArgoApps:      true,
+		},
+		{
+			Name: "two applications in the overview but none is updated",
+			Steps: []step{
+				{
+					RecvErr:       status.Error(codes.Canceled, "context cancelled"),
+					CancelContext: true,
+				},
+			},
+			Overview: &api.GetOverviewResponse{
+				Applications: map[string]*api.Application{
+					"foo": {
+						Releases: []*api.Release{
+							{
+								Version:        1,
+								SourceCommitId: "00001",
+							},
+						},
+						Team: "footeam",
+					},
+					"foo2": {
+						Releases: []*api.Release{
+							{
+								Version:        1,
+								SourceCommitId: "00012",
+							},
+						},
+						Team: "footeam",
+					},
+				},
+				EnvironmentGroups: []*api.EnvironmentGroup{
+					{
+
+						EnvironmentGroupName: "staging-group",
+						Environments: []*api.Environment{
+							{
+								Name: "staging",
+								Applications: map[string]*api.Environment_Application{
+									"foo": {
+										Name:    "foo",
+										Version: 1,
+										DeploymentMetaData: &api.Environment_Application_DeploymentMetaData{
+											DeployTime: "123456789",
+										},
+									},
+									"foo2": {
+										Name:    "foo2",
+										Version: 1,
+										DeploymentMetaData: &api.Environment_Application_DeploymentMetaData{
+											DeployTime: "1234567892",
+										},
+									},
+								},
+								Priority: api.Priority_UPSTREAM,
+								Config: &api.EnvironmentConfig{
+									Argocd: &api.EnvironmentConfig_ArgoCD{
+										Destination: &api.EnvironmentConfig_ArgoCD_Destination{
+											Name:   "staging",
+											Server: "test-server",
+										},
+									},
+								},
+							},
+						},
+					},
+				},
+				GitRevision: "1234",
+			},
+			ExpectedConsumed:      0,
+			ExpectedConsumedTypes: []string{},
+			ExistingArgoApps:      true,
 		},
 	}
 	for _, tc := range tcs {
@@ -640,7 +679,7 @@ func TestArgoConsume(t *testing.T) {
 			hlth.BackOffFactory = func() backoff.BackOff { return backoff.NewConstantBackOff(0) }
 			errCh := make(chan error)
 			go func() {
-				errCh <- argoProcessor.Consume(t, ctx, tc.ExpectedConsumedTypes)
+				errCh <- argoProcessor.Consume(t, ctx, tc.ExpectedConsumedTypes, tc.ExistingArgoApps)
 			}()
 
 			go func() {
@@ -684,19 +723,16 @@ func (a mockArgoProcessor) DeleteArgoApps(ctx context.Context, appsKnownToArgo m
 }
 
 func (a mockArgoProcessor) CreateOrUpdateApp(ctx context.Context, overview *api.GetOverviewResponse, app *api.Environment_Application, env *api.Environment, appsKnownToArgo map[string]*v1alpha1.Application) {
-	k := Key{AppName: app.Name, EnvName: env.Name, Application: app, Environment: env}
-
-	appExists := false
-
+	var existingApp *v1alpha1.Application
 	for _, argoApp := range appsKnownToArgo {
-		if argoApp.Name == env.Name && argoApp.Annotations["com.freiheit.kuberpult/application"] != "" {
-			appExists = true
+		if argoApp.Name == fmt.Sprintf("%s-%s", env.Name, app.Name) && argoApp.Annotations["com.freiheit.kuberpult/application"] != "" {
+			existingApp = argoApp
 			break
 		}
 	}
 
-	if !appExists {
-		appToCreate := CreateArgoApplication(overview, app, k.Environment)
+	if existingApp == nil {
+		appToCreate := CreateArgoApplication(overview, app, env)
 		appToCreate.ResourceVersion = ""
 		upsert := false
 		validate := false
@@ -713,13 +749,19 @@ func (a mockArgoProcessor) CreateOrUpdateApp(ctx context.Context, overview *api.
 			}
 		}
 	} else {
-		appToUpdate := CreateArgoApplication(overview, app, k.Environment)
-		appUpdateRequest := &application.ApplicationUpdateSpecRequest{
-			Name:         &appToUpdate.Name,
-			Spec:         &appToUpdate.Spec,
-			AppNamespace: &appToUpdate.Namespace,
+		appToUpdate := CreateArgoApplication(overview, app, env)
+		appUpdateRequest := &application.ApplicationUpdateRequest{
+			XXX_NoUnkeyedLiteral: struct{}{},
+			XXX_unrecognized:     nil,
+			XXX_sizecache:        0,
+			Validate:             ptr.Bool(false),
+			Application:          appToUpdate,
+			Project:              ptr.FromString(appToUpdate.Spec.Project),
 		}
-		a.ApplicationClient.UpdateSpec(ctx, appUpdateRequest)
+
+		if !cmp.Equal(appUpdateRequest.Application.Spec, existingApp.Spec, cmp.AllowUnexported(v1alpha1.ApplicationSpec{}.Destination)) {
+			a.ApplicationClient.Update(ctx, appUpdateRequest)
+		}
 	}
 }
